@@ -1,25 +1,31 @@
 {
-  "name": "app",
+  "name": "tarinapeli-react",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "final-form": "^4.10.0",
-    "prop-types": "^15.6.2",
-    "react": "^15.6.2",
-    "react-dom": "^15.6.2",
-    "react-final-form": "^3.6.5",
-    "react-redux": "^5.0.7",
-    "redux": "^4.0.0",
+    "final-form": "^4.12.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-final-form": "^4.1.0",
+    "react-redux": "^7.0.2",
+    "redux": "^4.0.1",
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
-    "react-scripts": "0.9.5"
+    "react-scripts": "2.1.8"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "proxy": "http://localhost:9080"
+  "proxy": "http://localhost:9080",
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
